add check if fq already aligned

+++ ChipSeq/alignment.sh
@@ -15,24 +15,28 @@ do
   foldername=${foldername%/*}
   basename=${basename##*/}
   
-  echo ${DATADIR}/${f}
 
-  mkdir -p ${foldername}
-  bowtie2 -x ${REFGENOME}/genome -U ${f} -S ${foldername}/${basename}.sam &> ${foldername}/${basename}.bowtie.log
- 
+  if [-f ${foldername}/${basename}_depDup_q30.bam]
+  then
+    else  
 
-  ## convert to sam files
-  samtools view -bSo ${foldername}/${basename}.bam ${foldername}/${basename}.sam
-  samtools sort ${foldername}/${basename}.bam > ${foldername}/${basename}_sorted.bam
-  ## index bam files for QC package
-  samtools index ${foldername}/${basename}_sorted.bam
-  rm ${foldername}/${basename}.sam
+	  mkdir -p ${foldername}
+	  bowtie2 -x ${REFGENOME}/genome -U ${f} -S ${foldername}/${basename}.sam &> ${foldername}/${basename}.bowtie.log
+	 
 
-  ## remove duplicates
-  java -jar $EBROOTPICARD/picard.jar MarkDuplicates I=${foldername}/${basename}_sorted.bam O=${foldername}/${basename}_depDuplicated.bam M=${foldername}/${basename}_dupMetrics.txt REMOVE_DUPLICATES=TRUE
-  
-  ## remove reads with q < 30 nb 
-  samtools view -q 30 -h ${foldername}/${basename}_depDuplicated.bam > ${foldername}/${basename}_depDup_q30.bam
+	  ## convert to sam files
+	  samtools view -bSo ${foldername}/${basename}.bam ${foldername}/${basename}.sam
+	  samtools sort ${foldername}/${basename}.bam > ${foldername}/${basename}_sorted.bam
+	  ## index bam files for QC package
+	  samtools index ${foldername}/${basename}_sorted.bam
+	  rm ${foldername}/${basename}.sam
+
+	  ## remove duplicates
+	  java -jar $EBROOTPICARD/picard.jar MarkDuplicates I=${foldername}/${basename}_sorted.bam O=${foldername}/${basename}_depDuplicated.bam M=${foldername}/${basename}_dupMetrics.txt REMOVE_DUPLICATES=TRUE
+	  
+	  ## remove reads with q < 30 nb 
+	  samtools view -q 30 -h ${foldername}/${basename}_depDuplicated.bam > ${foldername}/${basename}_depDup_q30.bam
+  fi	 
    
 done
